find_additional_works_from_input_csv/query_db/requirements.txt
duckdb==1.3.1
pyyaml==6.0.2
unidecode==1.4.0
pandas==2.2.3
jellyfish==1.1.0
nameparser==1.1.3
rapidfuzz==3.10.1
flair==0.14.0
